margi-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
google-generativeai>=0.7.2
resend==0.7.0
razorpay==1.4.1
uvloop==0.19.0